perf(chromatic): exclude demo-recorder stories from CI snapshots
The DemoFlow and FullDemo stories exist to drive the homepage demo
recorder; their play() functions burn ~30-50s of Chromatic wall time
even though both have disableSnapshot. Move them into App.demo.stories.tsx
and skip *.demo.stories.* from the Storybook glob when SKIP_DEMO_STORIES
is set, which the chromatic workflow now does. Local Storybook and
record-demo keep loading them.
